langgraph-agent/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
langgraph>=0.0.40
langchain-core>=0.1.0
dashscope>=1.14.0
python-dotenv>=1.0.0
langsmith>=0.0.80
requests>=2.31.0
